+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added method `query.WithIssuesHandler` to get query issues
+
 ## v3.117.2
 * Added support for `Result.RowsAffected()` for YDB `database/sql` driver
 * Upgraded minimal version of Go to 1.23.9